fix dgc compressor return types
